# Conflicts:
#	Senior Project/Assets/Animations/BossAnimations/Attacks_v3/Melee_left.anim
#	Senior Project/Assets/Animations/BossAnimations/Attacks_v3/Melee_right.anim
#	Senior Project/Assets/Animations/BossAnimations/Attacks_v3/Throw_left.anim
#	Senior Project/Assets/Animations/BossAnimations/Attacks_v3/Throw_right.anim
#	Senior Project/Assets/Prefabs/JeffTheBoss.prefab
#	Senior Project/Assets/Scenes/_Dead Scenes/AI_Test.unity
#	Senior Project/Assets/Scripts/EnemyScripts/BossAttacks/AttackMotion.cs
#	Senior Project/Assets/Scripts/EnemyScripts/BossAttacks/AttacksManager.cs
#	Senior Project/Assets/Scripts/EnemyScripts/BossAttacks/FireMeleeAttacks.cs
#	Senior Project/Assets/Scripts/EnemyScripts/BossAttacks/FireProjectileAttacks.cs
#	Senior Project/Assets/Scripts/EnemyScripts/BossAttacks/IceMeleeAttacks.cs
#	Senior Project/Assets/Scripts/EnemyScripts/BossAttacks/IceProjectileAttacks.cs
#	Senior Project/Assets/Scripts/EnemyScripts/BossEnemyController.cs

+++ Senior Project/Assets/Scripts/EnemyScripts/BossAttacks/AttackMotion.cs
@@ -56,7 +56,7 @@ public class AttackMotion : ScriptableObject
         yield return null;
     }
 
-    public virtual IEnumerator AttackingPlayer(float i)
+    public virtual IEnumerator AttackingPlayer(int i)
     {
         yield return null;
     }
